+++ include/common/monolish_coo.hpp
@@ -23,6 +23,7 @@
 
 namespace monolish {
 template <typename Float> class vector;
+template <typename TYPE, typename Float> class view1D;
 namespace matrix {
 template <typename Float> class Dense;
 template <typename Float> class CRS;
@@ -559,28 +560,6 @@ public:
   std::string type() const { return "COO"; }
 
   /**
-   * @brief get row vector
-   * @param i row number
-   * @param vec row vector
-   * @note
-   * - # of computation: nnz
-   * - Multi-threading: false
-   * - GPU acceleration: false
-   **/
-  void row(std::size_t i, vector<Float> &vec) const;
-
-  /**
-   * @brief get column vector
-   * @param i column number
-   * @param vec column vector
-   * @note
-   * - # of computation: nnz
-   * - Multi-threading: false
-   * - GPU acceleration: false
-   **/
-  void col(std::size_t i, vector<Float> &vec) const;
-
-  /**
    * @brief get diag. vector
    * @param vec diag. vector
    * @note
@@ -589,6 +568,34 @@ public:
    * - GPU acceleration: false
    **/
   void diag(vector<Float> &vec) const;
+  void diag(view1D<vector<Float>, Float> &vec) const;
+  void diag(view1D<matrix::Dense<Float>, Float> &vec) const;
+
+  /**
+   * @brief get row vector
+   * @param r row number
+   * @param vec row vector
+   * @note
+   * - # of computation: nnz
+   * - Multi-threading: false
+   * - GPU acceleration: false
+   **/
+  void row(const size_t r, vector<Float> &vec) const;
+  void row(const size_t r, view1D<vector<Float>, Float> &vec) const;
+  void row(const size_t r, view1D<matrix::Dense<Float>, Float> &vec) const;
+
+  /**
+   * @brief get column vector
+   * @param c column number
+   * @param vec column vector
+   * @note
+   * - # of computation: nnz
+   * - Multi-threading: false
+   * - GPU acceleration: false
+   **/
+  void col(const size_t c, vector<Float> &vec) const;
+  void col(const size_t c, view1D<vector<Float>, Float> &vec) const;
+  void col(const size_t c, view1D<matrix::Dense<Float>, Float> &vec) const;
 
   /////////////////////////////////////////////////////////////////////////////