+++ include/common/monolish_crs.hpp
@@ -15,6 +15,7 @@
 
 namespace monolish {
 template <typename Float> class vector;
+template <typename TYPE, typename Float> class view1D;
 namespace matrix {
 template <typename Float> class Dense;
 template <typename Float> class COO;
@@ -344,6 +345,8 @@ public:
    * - GPU acceleration: true
    **/
   void diag(vector<Float> &vec) const;
+  void diag(view1D<vector<Float>, Float> &vec) const;
+  void diag(view1D<matrix::Dense<Float>, Float> &vec) const;
 
   /**
    * @brief get row vector
@@ -355,6 +358,8 @@ public:
    * - GPU acceleration: true
    **/
   void row(const size_t r, vector<Float> &vec) const;
+  void row(const size_t r, view1D<vector<Float>, Float> &vec) const;
+  void row(const size_t r, view1D<matrix::Dense<Float>, Float> &vec) const;
 
   /**
    * @brief get column vector
@@ -366,6 +371,8 @@ public:
    * - GPU acceleration: true
    **/
   void col(const size_t c, vector<Float> &vec) const;
+  void col(const size_t c, view1D<vector<Float>, Float> &vec) const;
+  void col(const size_t c, view1D<matrix::Dense<Float>, Float> &vec) const;
 
   /////////////////////////////////////////////////////////////////////////////