+++ include/common/monolish_dense.hpp
@@ -10,6 +10,7 @@
 
 namespace monolish {
 template <typename Float> class vector;
+template <typename TYPE, typename Float> class view1D;
 namespace matrix {
 /**
  * @brief Dense format Matrix
@@ -398,6 +399,8 @@ public:
    * - GPU acceleration: true
    **/
   void diag(vector<Float> &vec) const;
+  void diag(view1D<vector<Float>, Float> &vec) const;
+  void diag(view1D<matrix::Dense<Float>, Float> &vec) const;
 
   /**
    * @brief get row vector
@@ -409,6 +412,8 @@ public:
    * - GPU acceleration: true
    **/
   void row(const size_t r, vector<Float> &vec) const;
+  void row(const size_t r, view1D<vector<Float>, Float> &vec) const;
+  void row(const size_t r, view1D<matrix::Dense<Float>, Float> &vec) const;
 
   /**
    * @brief get column vector
@@ -420,6 +425,8 @@ public:
    * - GPU acceleration: true
    **/
   void col(const size_t c, vector<Float> &vec) const;
+  void col(const size_t c, view1D<vector<Float>, Float> &vec) const;
+  void col(const size_t c, view1D<matrix::Dense<Float>, Float> &vec) const;
 
   /////////////////////////////////////////////////////////////////////////////